content_analyzer/requirements.txt
google-cloud-language
newspaper3k
pymongo
